+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/IssuerEncryptionService.kt
@@ -5,16 +5,15 @@ import at.asitplus.catching
 import at.asitplus.openid.CredentialRequestParameters
 import at.asitplus.openid.CredentialResponseParameters
 import at.asitplus.openid.SupportedAlgorithmsContainer
-import at.asitplus.signum.indispensable.josef.JsonWebKey
 import at.asitplus.signum.indispensable.josef.JsonWebKeySet
 import at.asitplus.signum.indispensable.josef.JweAlgorithm
 import at.asitplus.signum.indispensable.josef.JweEncrypted
 import at.asitplus.signum.indispensable.josef.JweEncryption
 import at.asitplus.signum.indispensable.josef.JweHeader
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
-import at.asitplus.signum.indispensable.josef.toJsonWebKey
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.KeyMaterial
+import at.asitplus.wallet.lib.agent.toEncryptionJsonWebKey
 import at.asitplus.wallet.lib.jws.DecryptJwe
 import at.asitplus.wallet.lib.jws.DecryptJweFun
 import at.asitplus.wallet.lib.jws.EncryptJwe
@@ -56,7 +55,7 @@ class IssuerEncryptionService @JvmOverloads constructor(
             supportedEncryptionAlgorithmsStrings = supportedJweEncryptionAlgorithms.map { it.identifier }.toSet(),
             encryptionRequired = requireRequestEncryption || requireResponseEncryption,
             jsonWebKeySet = JsonWebKeySet(
-                listOf(decryptionKeyMaterial.publicKey.toJsonWebKey(decryptionKeyMaterial.identifier).forEncryption())
+                listOf(decryptionKeyMaterial.toEncryptionJsonWebKey())
             )
         )
     else null
@@ -140,8 +139,4 @@ class IssuerEncryptionService @JvmOverloads constructor(
                 throw InvalidEncryptionParameters("Response encryption required, no params sent")
             CredentialIssuer.CredentialResponse.Plain(response)
         }
-
-    // should always be ecdh-es for encryption
-    private fun JsonWebKey.forEncryption(): JsonWebKey =
-        this.copy(algorithm = JweAlgorithm.ECDH_ES, publicKeyUse = "enc")
 }